Always include LineageOS sepolicy and config.mk
Some essential LineageOS configs are only included if LINEAGE_BUILD var
is set, which is only the case for products that begin with "lineage_".
To support "maru_" products we must make sure to include these. There's
really no point in not including them anyways, since we are never doing
standard AOSP builds on a LineageOS tree/manifest.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>