Fix missing resources for maru products
There are some resources (sepolicy for example) that are only included
for builds in which the `LINEAGE_BUILD` environment variable is set.
LINEAGE_BUILD is only set for products that being with `lineage_` so it
causes our maru products to break. Let's just make sure this variable is
set for maru products as well.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>